1/23/2019
##Refactored
1) Moved the Login method from the LoginFragment to the LoginAct. By doing this, both the LoginFrag and the SignUpFrag can now call the same Login (i.e. the emailLogIn method) via the LoginAct.

ToDo:
There is a strange interaction when navigating from the home screen back into the app; there is a short animation where the Main Activity starts but quickly transitions into the Login Activity. This should be the case when we return to the app and there is no user logged in but not sure how to resolve.

##Login Activity Progress:
1) Migrated Firebase signup to the Signup Fragment.
2) Button navigation between Login and Signup fragments successfully bound.
3) Signup Fragment checks all fields are properly filled before registering user.
--SignUp Fragment seems to be fully functional.
4) Login fragment properly authorizes users.
5) Main activity is launched on successful login.
6) Pressing the back button in Login Activity brings the user to home.
7) Pressing back on the MainActivity brings user to home (currently untested for other Activities in the back stack.