1/23/2019 - Implementing Room Library
ToDo:
There is a strange interaction when navigating from the home screen back into the app; there is a short animation where the Main Activity starts but quickly transitions into the Login Activity. This should be the case when we return to the app and there is no user logged in but not sure how to resolve.

##Room Implementation Progress:
1) Redefined the previous LiftObject class to an EntityObject class.