1/24/2019 - Implementing Room Library and RecyclerView
-RecyclerView components created but untested. Not exactly sure how to connect the Room Database to the RecyclerView.

ToDo:
There is a strange interaction when navigating from the home screen back into the app; there is a short animation where the Main Activity starts but quickly transitions into the Login Activity. This should be the case when we return to the app and there is no user logged in but not sure how to resolve.

##Room Implementation Progress:
1) Redefined the previous LiftObject class to an EntityObject class.
2) Created ViewModel, Repository, Room Database, and DAO for StatEntity.