2/5/2019
Figure out how to display the child detail fragment on top of our other fragment.
Add RecyclerView onClick implementation to the Events CardViews.

OnClick starts a new Fragment that passes the GroupEvent that was clicked to the Fragment that is then populated with the event's data into a new view 'event_detail_frag'.