2/8/2019
Able to populate the EventRecyclerView when the Fragment starts.

The CreateEventFragment creates a GroupEvent and stores a tag under the User's Events Document as well as stores a GroupEvent POJO in the Events collection.

When on the screen where user  views their groups, clicking on a card will inflate a new fragment for the user to see the details of the group.

Need to add a method to see if the user is an admin or owner, if yes allow the user to edit the contents of the group.