2/12/2019
Added Splash Screen when cold starting the app. This launcher activity will also implement the FirebaseAuth listener and determine whether to start the main_activity or the login_activity.

The reason app was crashing when user launched into the GroupsActivity is because the old User documents contained Group names of objects that were not serializable as GroupBase objects back to the application.