2/17/2019
AlertDialog launches the camera app if the 'Camera' imageView is clicked.

Need to handle the image preview by prompting user if they want to use this image or retake the photo.

ToDo:
Open gallery when galleryView is clicked. One item selected and confirmed, we want to load the Group's imageView with that selected photo. On confirm, check to see if photo is selected if yes store it to FireStore.